% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melsvmp.R
\name{mels_vmp}
\alias{mels_vmp}
\title{Fit a Mixed-Effects Location Scale Model}
\usage{
mels_vmp(y, beta_formula, alpha_formula, tau_formula, id, data, ...)
}
\arguments{
\item{y}{The name of the response variable in 'data'.}

\item{beta_formula}{A formula for the mean model (fixed effects).}

\item{alpha_formula}{A formula for the between-subject variance.}

\item{tau_formula}{A formula for the within-subject variance.}

\item{id}{The name of the subject ID variable in 'data'.}

\item{data}{A data.frame containing all variables.}

\item{...}{Additional arguments passed to the fitter function.}
}
\value{
An object of class 'mels_vmp'.
}
\description{
Uses a Variational Message Passing (VMP) algorithm.
}
