% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{riesby}
\alias{riesby}
\title{Riesby Data Set}
\format{
A data frame with 396 rows and 6 variables:
\describe{
\item{id}{Subject ID}
\item{hamd}{Hamilton depression score}
\item{intcpt}{Intercept term}
\item{week}{Week of study}
\item{endog}{Endogenous depression status}
\item{endweek}{Interaction of endog and week}
}
}
\usage{
riesby
}
\description{
A longitudinal dataset that examined the Hamilton depression plasma scores and
clinical response
}
\details{
66 depressed inpatients (37 endogenous and 29 non-endogenous) with 396 total
observations.
}
\references{
Reisby, Niels, et al. "Imipramine: clinical effects and pharmacokinetic variability." Psychopharmacology 54.3 (1977): 263-272.
}
\keyword{datasets}
