% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melsvmp.R
\name{bootstrap_mels_vmp}
\alias{bootstrap_mels_vmp}
\title{Bootstrap method for mels_vmp objects}
\usage{
bootstrap_mels_vmp(model_object, B = 1000, parallel = TRUE, cores = 5, ...)
}
\arguments{
\item{model_object}{An object of class 'mels_vmp'.}

\item{B}{Number of bootstrap replicates.}

\item{parallel}{Logical, whether to run in parallel. (ADD THIS)}

\item{cores}{Number of cores for parallel execution. (ADD THIS)}

\item{...}{Other arguments. (This line is fine)}
}
\description{
Bootstrap method for mels_vmp objects
}
