% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{riesby}
\alias{riesby}
\title{Riesby Data Set
...}
\format{
A data frame with X rows and 6 variables:
\describe{
\item{id}{Subject ID}
\item{hamd}{Hamilton depression score}
\item{intcpt}{Intercept term}
\item{week}{Week of study}
\item{endog}{Endogenous depression status}
\item{endweek}{Interaction of endog and week}
}
}
\usage{
riesby
}
\description{
Riesby Data Set
...
}
\keyword{datasets}
