% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melsvmp.R
\name{summary.mels_vmp}
\alias{summary.mels_vmp}
\title{Summary method for mels_vmp objects}
\usage{
\method{summary}{mels_vmp}(object, ...)
}
\arguments{
\item{object}{An object of class 'mels_vmp'.}

\item{...}{Additional arguments}
}
\description{
Provides a summary of the fitted model with robust sandwich errors.
}
