% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melsvmp.R
\name{summary.mels_vmp_bootstrap}
\alias{summary.mels_vmp_bootstrap}
\title{Summary method for mels_vmp_bootstrap objects}
\usage{
\method{summary}{mels_vmp_bootstrap}(object, ...)
}
\arguments{
\item{object}{An object of class 'mels_vmp_bootstrap'.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Provides a summary of the bootstrap results.
}
